moves Footer into its own component file - App imports Footer
